Revert "Merge pull request #134 from grab/resanatize-the-lint-baseline"
This reverts commit 3f16c52ebc93571ad8e8b0ca1ef1e65713d257ea, reversing
changes made to 4db9d71787fcfd546412382c4cdcc2ad25d7992e.
