+++ tools/lint/src/main/java/com/grab/lint/LintAnalyzeCommand.kt
@@ -1,13 +1,5 @@
 package com.grab.lint
 
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.asFlow
-import kotlinx.coroutines.flow.collect
-import kotlinx.coroutines.flow.flatMapMerge
-import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.runBlocking
 import java.io.File
 import java.nio.file.Path
 import kotlin.io.path.pathString
@@ -21,6 +13,7 @@ class LintAnalyzeCommand : LintBaseCommand() {
         workingDir: Path,
         projectXml: File,
         tmpBaseline: File,
+        lintBaselineHandler: LintBaselineHandler
     ) {
         val cliArgs = (defaultLintOptions + listOf(
             "--cache-dir", workingDir.resolve("cache").pathString,
@@ -28,16 +21,5 @@ class LintAnalyzeCommand : LintBaseCommand() {
             "--analyze-only" // Only do analyze
         )).toTypedArray()
         LintCli().run(cliArgs)
-        sanitizePartialResults()
-    }
-
-    private fun sanitizePartialResults() = runBlocking {
-        val concurrency = Runtime.getRuntime().availableProcessors() / 2
-        partialResults.walkTopDown().filter { it.isFile }.asFlow()
-        .flatMapMerge(concurrency) { partialResult ->
-            flow { emit(partialResult) }
-                .map { sanitizer.sanitize(it) }
-                .flowOn(Dispatchers.IO)
-        }.collect()
     }
 }
